backend/requirements.txt
fastapi
uvicorn
PyJWT
bcrypt
google-genai
mistralai
python-dotenv